mass commit (very bad, sorry)
git-svn-id: https://svn.salilab.org/imp/trunk@11483 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/isd/include/MarginalHBondRestraint.h
@@ -55,7 +55,7 @@ public:
   //return the sum inside the parentheses
   double get_logsquares() const {return logsquares_;}
 
-  double get_number_of_contributions() const {return volumes_.size();}
+  unsigned get_number_of_contributions() const {return volumes_.size();}
 
   /* call for probability */
   double get_probability() const
