+++ modules/isd/include/MarginalNOERestraint.h
@@ -59,7 +59,7 @@ public:
   //return the sum of squares wrt current structure.
   double get_SS() const {return SS_;}
 
-  double get_number_of_contributions() const {return volumes_.size();}
+  unsigned get_number_of_contributions() const {return volumes_.size();}
 
   /* call for probability */
   double get_probability() const
